compose-requirements.txt
python-dotenv==1.*
pydantic==2.*